requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
requests
google-cloud-logging
jwt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
boto3